Versuchsaufbau im Labor-> Parameter angepasst, Pi sollte eine Auflösung von 1920x1080 liefern damit man gute Werte bekommt

+++ qtProjekt/ServerGui/model/PositionService/LocatableObject.cpp
@@ -10,7 +10,7 @@ LocatableObject::LocatableObject(){
 
 }
 
-LocatableObject::LocatableObject(Form _form, int _colorMin, int _colorMax, float _sizeInCm) {
+LocatableObject::LocatableObject(Form _form, int _colorMin, int _colorMax, double _sizeInCm) {
 	
 	form = _form;
 	colorMin = _colorMin;
